Application to raw sql (#735)
* Normalize DB in terms of roles and tags

Roles now exist in the db in their own table.
Tags no longer use an array for their versions.

* Mostly done with moving application to doobie

* Fix bugs

* Try to fix Travis

* Decouple role ids from forum role ids

* Fix view

* Stupid red code

* Use a unique index

* IO can be converted to Future

* Scalafmt

* Scalafmt

* Scalafmt

* More SQL queries

* User trust views for getting the trust in general

* Create trust views evolution

* Some minor updates. Home projects is now a materialized view

* Most stuff works

* Add (broken) tests

* Fix tests and some other stuff, use home_projects view for user projects

* Optimizations and fixes

* Even more fixes. Project approvals working now

* Use postgres 11.1 for travis

* Maybe fix travis?

* Fix build?

* Try postgres 11

* Use travis user and db. Set password for travis user

* Try with 10 and see what happens

* Try xenial

* Can't use oraclejdk8

* Don't left join on roles

* Some small fixes

* Some more fixes. Update the view in some cases

* Show project even if no recommended version was found

+++ app/controllers/ApiController.scala
@@ -78,8 +78,8 @@ final class ApiController @Inject()(
       categories: Option[String],
       sort: Option[Int],
       q: Option[String],
-      limit: Option[Int],
-      offset: Option[Int]
+      limit: Option[Long],
+      offset: Option[Long]
   ): Action[AnyContent] = Action.asyncF {
     version match {
       case "v1" => this.api.getProjectList(categories, sort, q, limit, offset).map(Ok(_))
